requirements.txt
# requirements.txt

selenium-wire
beautifulsoup4
requests